[Refactor] style.ts 파일 위치 이동 #261
